moved machinestate contents back to machine.py
